+++ runtime/Clib/cprocess.c
@@ -58,7 +58,7 @@ extern long  bgl_list_length( obj_t );
 extern char *bgl_string_to_gc_cstring( obj_t );
 extern obj_t string_to_bstring( char * );
 extern ssize_t bgl_syswrite( obj_t, char *, size_t );
-extern obj_t make_string_sans_fill( int );
+extern obj_t make_string_sans_fill( long );
 
 /*---------------------------------------------------------------------*/
 /*    Prototypes                                                       */
